+++ stacierl/replaybuffer/vanillalstm.py
@@ -62,12 +62,12 @@ class VanillaLSTM(ReplayBuffer):
                 n_padding = self.sequence_length - len(episode)
 
                 state_padding_shape = tuple([n_padding]) + state_seq[0].shape
-                state_padding = np.full(state_padding_shape, 0.0, dtype=np.float16)
+                state_padding = np.full(state_padding_shape, 0.0, dtype=np.float32)
                 state_seq = np.concatenate((state_padding, state_seq), axis=0)
                 next_state_seq = np.concatenate((state_padding, next_state_seq), axis=0)
 
                 action_padding_shape = tuple([n_padding]) + action_seq[0].shape
-                action_padding = np.full(action_padding_shape, 0.0, dtype=np.float16)
+                action_padding = np.full(action_padding_shape, 0.0, dtype=np.float32)
                 action_seq = np.concatenate((action_padding, action_seq), axis=0)
 
                 reward_seq = [0] * n_padding + reward_seq
